+++ clients/admin-ui/src/types/api/models/PrivacyNoticeRegion.ts
@@ -1,0 +1,40 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * Enum is not formalized in the DB because it is subject to frequent change
+ */
+export enum PrivacyNoticeRegion {
+  US_CA = "us_ca",
+  US_CO = "us_co",
+  US_VA = "us_va",
+  US_UT = "us_ut",
+  EU_BE = "eu_be",
+  EU_BG = "eu_bg",
+  EU_CZ = "eu_cz",
+  EU_DK = "eu_dk",
+  EU_DE = "eu_de",
+  EU_EE = "eu_ee",
+  EU_IE = "eu_ie",
+  EU_EL = "eu_el",
+  EU_ES = "eu_es",
+  EU_FR = "eu_fr",
+  EU_HR = "eu_hr",
+  EU_IT = "eu_it",
+  EU_CY = "eu_cy",
+  EU_LV = "eu_lv",
+  EU_LT = "eu_lt",
+  EU_LU = "eu_lu",
+  EU_HU = "eu_hu",
+  EU_MT = "eu_mt",
+  EU_NL = "eu_nl",
+  EU_AT = "eu_at",
+  EU_PL = "eu_pl",
+  EU_PT = "eu_pt",
+  EU_RO = "eu_ro",
+  EU_SI = "eu_si",
+  EU_SK = "eu_sk",
+  EU_FI = "eu_fi",
+  EU_SE = "eu_se",
+}